add dockerfile into template
