+++ tools/goctl/tpl/templates.go
@@ -6,6 +6,7 @@ import (
 	"github.com/logrusorgru/aurora"
 	"github.com/tal-tech/go-zero/core/errorx"
 	"github.com/tal-tech/go-zero/tools/goctl/api/gogen"
+	"github.com/tal-tech/go-zero/tools/goctl/docker"
 	modelgen "github.com/tal-tech/go-zero/tools/goctl/model/sql/gen"
 	rpcgen "github.com/tal-tech/go-zero/tools/goctl/rpc/generator"
 	"github.com/tal-tech/go-zero/tools/goctl/util"
@@ -24,6 +25,9 @@ func GenTemplates(ctx *cli.Context) error {
 		},
 		func() error {
 			return rpcgen.GenTemplates(ctx)
+		},
+		func() error {
+			return docker.GenTemplates(ctx)
 		},
 	); err != nil {
 		return err
